@@ -1,6 +1,6 @@
 module websocket_client_chat
 
-go 1.19
+go 1.21
 
 require (
 	github.com/gordonklaus/portaudio v0.0.0-20250206071425-98a94950218b
